# guest-role/WebContent/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "菜单角色调试接口",
    "version" : "v1",
    "title" : "Guest Role",
    "termsOfService" : "http://www.github.com/kongchen/swagger-maven-plugin",
    "contact" : {
      "name" : "Luo Jing",
      "email" : "luojing@zhiweicloud.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "192.168.1.126:8080",
...
  "schemes" : [ "http" ],
  "paths" : {
    "/addMenu" : {
      "post" : {
        "tags" : [ "菜单管理" ],
        "summary" : "菜单管理 - 新增/修改",
        "description" : "返回成功还是失败",
        "operationId" : "addMenu",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json;charset=utf-8" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "sysMenu",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/全局统一接受参数SysMenu"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
    "/deleteMenus" : {
      "post" : {
        "tags" : [ "菜单管理" ],
        "summary" : "菜单管理 - 删除",
        "description" : "返回响应结果",
        "operationId" : "deleteMenus",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json;charset=utf8" ],
        "parameters" : [ {
          "name" : "airportCode",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/全局统一接受参数Long"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "/getMenuByRoleId" : {
      "get" : {
        "tags" : [ "菜单管理" ],
        "summary" : "菜单管理 - 查询当前角色有哪些菜单",
        "description" : "返回成功还是失败",
        "operationId" : "getMenuByRoleId",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json;charset=utf-8" ],
        "parameters" : [ {
          "name" : "airportCode",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "roleId",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "/menuTree" : {
      "get" : {
        "tags" : [ "菜单管理" ],
        "summary" : "菜单管理 - 菜单树",
        "description" : "返回成功还是失败",
        "operationId" : "menuTree",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json;charset=utf-8" ],
        "parameters" : [ {
          "name" : "airportCode",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "请求成功",
            "schema" : {
              "$ref" : "#/definitions/SysMenu"
            }
          },
          "400" : {
            "description" : "参数错误"
          },
          "405" : {
            "description" : "请求方式不对"
          }
        }
      }
    },
...
    "/viewMenu" : {
      "get" : {
        "tags" : [ "菜单管理" ],
        "summary" : "菜单管理 - 查询菜单详情 ",
        "description" : "返回查询菜单详情",
        "operationId" : "viewMenu",
        "produces" : [ "application/json;charset=utf8" ],
        "parameters" : [ {
          "name" : "menuId",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "airportCode",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "menuId",
          "in" : "query",
          "description" : "员工id",
          "required" : true,
          "type" : "Long"
        }, {
          "name" : "airportCode",
          "in" : "query",
          "description" : "机场编号",
          "required" : true,
          "type" : "String"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "SysMenu" : {
      "type" : "object",
      "required" : [ "airportCode", "name" ],
      "properties" : {
        "createTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "创建时间"
        },
        "updateTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "修改时间"
        },
        "isDeleted" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "是否删除：默认为0，0：不删除，1：删除"
        },
        "airportCode" : {
          "type" : "string",
          "description" : "机场code"
        },
        "createUser" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "创建人"
        },
        "updateUser" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "修改人"
        },
        "menuId" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "菜单id"
        },
        "pid" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "父id"
        },
        "name" : {
          "type" : "string",
          "description" : "菜单名称"
        },
        "url" : {
          "type" : "string",
          "description" : "菜单链接地址"
        },
        "type" : {
          "type" : "string",
          "description" : "菜单类型"
        },
        "position" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "菜单排序号"
        },
        "children" : {
          "type" : "array",
          "description" : "子菜单",
          "items" : {
            "$ref" : "#/definitions/SysMenu"
          }
        }
      },
      "description" : "sys_menu"
    },
...
    "全局统一接受参数Long" : {
      "type" : "object",
      "properties" : {
        "data" : {
          "type" : "array",
          "description" : "数据",
          "items" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "description" : "全局统一接受参数"
    },
    "全局统一接受参数SysMenu" : {
      "type" : "object",
      "properties" : {
        "data" : {
          "type" : "array",
          "description" : "数据",
          "items" : {
            "$ref" : "#/definitions/SysMenu"
          }
        }
      },
      "description" : "全局统一接受参数"
    },
...
  }
}